Move `Annotations` into the `model-compiler`
... to avoid the dependency from Java Poet in `base` module.
