+++ tools/model-compiler/src/main/java/io/spine/tools/compiler/validation/ValidatingBuilderWriter.java
@@ -38,7 +38,7 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 
-import static io.spine.code.java.Annotations.generatedBySpineModelCompiler;
+import static io.spine.tools.compiler.annotation.Annotations.generatedBySpineModelCompiler;
 import static io.spine.tools.compiler.validation.ClassNames.getValidatorMessageClassName;
 import static io.spine.util.Exceptions.newIllegalArgumentException;
