+++ tools/protoc-plugin/build.gradle
@@ -24,6 +24,7 @@ group 'io.spine.tools'
 
 dependencies {
     implementation project(':base')
+    implementation project(':model-compiler')
 
     implementation deps.gen.javaPoet
     testProtobuf project(':base')
